Second stage. Change the architecture, use Mapperly to map DTOs, finished 1st version of transaction core, standarize all classes in layers

+++ Accounting.Domain/DTO/CountryDTO.cs
@@ -2,7 +2,7 @@
 {
     public class CountryDTO
     {
-        public int CountryID { get; set; }
+        public short CountryID { get; set; }
         public string Name { get; set; } = string.Empty;
         public short? OfficialCoinID { get; set; }
         public CoinDTO? OfficialCoin { get; set; }
